lasertag: fix OPCODE_HIT to not use overlapping bit fields.
We had defined OPCODE_HIT with a 16-bit payload with bits 4-13 as
the badge ID of the shooter.  However, we define bits 12-15 as
the opcode. So we are short 2 bits, as these fields overlap.
We were using bits 0-3 as the team.

To fix this, use bits 0-2 as the team (means we can only have 8
teams, not 16), and define bits 3-11 as badge id. and 12-15
as opcode.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>