lasertag: fix OPCODE_VENDOR_POWER_UP to avoid overlapping bit fields
We defined bits 4-13 as the badge ID of the vendor.  However, we
define bits 12-15 as the opcode. So we are short 2 bits. This was basically
copied from OPCODE_HIT (which was also broken).

To fix, we do not need any bits for the team as we did with OPCODE HIT, so
we can define the badge ID of the vendor to occupy bits 0-8 and this will
avoid overlapping the opcode bits (12-15).

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>